📦 refactor(resizable): fixing failed tests
