+++ apps/web/public/components/carousel/demo/sizes.md
@@ -1,8 +1,8 @@
 ```angular-ts showLineNumbers copyButton
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 
-import { ZardCardComponent } from '../../card';
-import { ZardCarouselImports } from '../carousel.imports';
+import { ZardCardComponent } from '@/shared/components/card';
+import { ZardCarouselImports } from '@/shared/components/carousel/carousel.imports';
 
 @Component({
   imports: [ZardCarouselImports, ZardCardComponent],
