+++ apps/web/public/components/resizable/demo/with-min-max.md
@@ -1,11 +1,11 @@
 ```angular-ts showLineNumbers copyButton
 import { Component } from '@angular/core';
 
-import { ResizableImports } from '@/shared/components/resizable/resizable.imports';
+import { ZardResizableImports } from '@/shared/components/resizable/resizable.imports';
 
 @Component({
   selector: 'z-demo-resizable-with-min-max',
-  imports: [ResizableImports],
+  imports: [ZardResizableImports],
   standalone: true,
   template: `
     <div class="space-y-4">
