+++ apps/web/public/components/select/demo/default.md
@@ -1,8 +1,8 @@
 ```angular-ts showLineNumbers copyButton
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 
-import { ZardBadgeComponent } from '../../badge/badge.component';
-import { ZardSelectImports } from '../select.imports';
+import { ZardBadgeComponent } from '@/shared/components/badge';
+import { ZardSelectImports } from '@/shared/components/select/select.imports';
 
 @Component({
   selector: 'z-demo-select-basic',
