+++ apps/web/public/components/sheet/demo/dimensions.md
@@ -1,8 +1,8 @@
 ```angular-ts showLineNumbers copyButton
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 
-import { ZardButtonComponent } from '../../button/button.component';
-import { ZardSheetService } from '../sheet.service';
+import { ZardButtonComponent } from '@/shared/components/button';
+import { ZardSheetService } from '@/shared/components/sheet/sheet.service';
 
 @Component({
   selector: 'z-demo-sheet-dimensions',
