+++ apps/web/public/components/table/demo/payments.md
@@ -1,5 +1,5 @@
 ```angular-ts showLineNumbers copyButton
-import { Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 import { ZardBadgeComponent } from '@/shared/components/badge/badge.component';
 import { ZardButtonComponent } from '@/shared/components/button/button.component';
@@ -16,7 +16,6 @@ export interface Payment {
 @Component({
   selector: 'z-demo-table-payments',
   imports: [ZardTableImports, ZardBadgeComponent, ZardButtonComponent, ZardIconComponent],
-  standalone: true,
   template: `
     <div class="w-full">
       <div class="overflow-hidden rounded-md border">
@@ -64,6 +63,7 @@ export interface Payment {
       </div>
     </div>
   `,
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ZardDemoTablePaymentsComponent {
   payments: Payment[] = [
