+++ apps/web/public/installation/manual/resizable.md
@@ -700,7 +700,7 @@ import { ZardResizableHandleComponent } from '@/shared/components/resizable/resi
 import { ZardResizablePanelComponent } from '@/shared/components/resizable/resizable-panel.component';
 import { ZardResizableComponent } from '@/shared/components/resizable/resizable.component';
 
-export const ResizableImports = [
+export const ZardResizableImports = [
   ZardResizableComponent,
   ZardResizableHandleComponent,
   ZardResizablePanelComponent,
