+++ libs/zard/src/lib/shared/components/carousel/index.ts
@@ -3,3 +3,4 @@ export * from '@/shared/components/carousel/carousel-content.component';
 export * from '@/shared/components/carousel/carousel-item.component';
 export * from '@/shared/components/carousel/carousel-plugins.service';
 export * from '@/shared/components/carousel/carousel.variants';
+export * from '@/shared/components/carousel/carousel.imports';
