+++ libs/zard/src/lib/shared/components/command/command-option-group.component.ts
@@ -17,7 +17,6 @@ import { mergeClasses } from '@/shared/utils/merge-classes';
 
 @Component({
   selector: 'z-command-option-group',
-  standalone: true,
   template: `
     @if (shouldShow()) {
       <div [class]="classes()" role="group">
