+++ libs/zard/src/lib/shared/components/layout/index.ts
@@ -4,3 +4,4 @@ export * from '@/shared/components/layout/footer.component';
 export * from '@/shared/components/layout/content.component';
 export * from '@/shared/components/layout/sidebar.component';
 export * from '@/shared/components/layout/layout.variants';
+export * from '@/shared/components/layout/layout.imports';
