+++ libs/zard/src/lib/shared/components/resizable/demo/vertical.ts
@@ -1,10 +1,10 @@
 import { Component } from '@angular/core';
 
-import { ResizableImports } from '@/shared/components/resizable/resizable.imports';
+import { ZardResizableImports } from '@/shared/components/resizable/resizable.imports';
 
 @Component({
   selector: 'z-demo-resizable-vertical',
-  imports: [ResizableImports],
+  imports: [ZardResizableImports],
   standalone: true,
   template: `
     <z-resizable zLayout="vertical" class="h-100 w-125 max-w-md rounded-lg border">
