+++ libs/zard/src/lib/shared/components/resizable/demo/with-min-max.ts
@@ -1,10 +1,10 @@
 import { Component } from '@angular/core';
 
-import { ResizableImports } from '@/shared/components/resizable/resizable.imports';
+import { ZardResizableImports } from '@/shared/components/resizable/resizable.imports';
 
 @Component({
   selector: 'z-demo-resizable-with-min-max',
-  imports: [ResizableImports],
+  imports: [ZardResizableImports],
   standalone: true,
   template: `
     <div class="space-y-4">
